Fixed QM scripts and SQL spawns for Abyssea-Konschtat
