+++ scripts/zones/Abyssea-Konschtat/npcs/qm12.lua
@@ -1,0 +1,48 @@
+-----------------------------------
+-- Zone: Abyssea-Konschtat
+--  NPC: qm12 (???)
+-- Spawns Fear Gorta
+-- @pos ? ? ? 15
+-----------------------------------
+require("scripts/globals/status");
+
+-----------------------------------
+-- onTrade Action
+-----------------------------------
+
+function onTrade(player,npc,trade)
+--[[
+    if (trade:hasItemQty(2905,1) and trade:getItemCount() == 1) then -- Player has all the required items.
+        if (GetMobAction(16839033) == ACTION_NONE) then -- Mob not already spawned from this
+            SpawnMob(16839033):updateClaim(player); -- Spawn NM, Despawn after inactive for 5 minutes (pt has to reclaim within 5 of a wipe)
+            player:tradeComplete();
+        end
+    end
+]]
+end;
+
+-----------------------------------
+-- onTrigger Action
+-----------------------------------
+
+function onTrigger(player,npc)
+    player:startEvent(1010, 2905); -- Inform payer what items they need.
+end;
+
+-----------------------------------
+-- onEventUpdate
+-----------------------------------
+
+function onEventUpdate(player,csid,option)
+    -- printf("CSID: %u",csid);
+    -- printf("RESULT: %u",option);
+end;
+
+-----------------------------------
+-- onEventFinish
+-----------------------------------
+
+function onEventFinish(player,csid,option)
+    -- printf("CSID: %u",csid);
+    -- printf("RESULT: %u",option);
+end;
