--- scripts/zones/Abyssea-Konschtat/npcs/qm20.lua
@@ -1,48 +1,0 @@
------------------------------------
--- Zone: Abyssea-Konschtat
---  NPC: qm20 (???)
--- Spawns Siranpa-Kamuy
--- @pos ? ? ? 15
------------------------------------
-require("scripts/globals/status");
-
------------------------------------
--- onTrade Action
------------------------------------
-
-function onTrade(player,npc,trade)
---[[
-    if (trade:hasItemQty(2906,1) and trade:getItemCount() == 1) then -- Player has all the required items.
-        if (GetMobAction(16838962) == ACTION_NONE) then -- Mob not already spawned from this
-            SpawnMob(16838962):updateClaim(player); -- Spawn NM, Despawn after inactive for 5 minutes (pt has to reclaim within 5 of a wipe)
-            player:tradeComplete();
-        end
-    end
-]]
-end;
-
------------------------------------
--- onTrigger Action
------------------------------------
-
-function onTrigger(player,npc)
-    player:startEvent(1010, 2906); -- Inform payer what items they need.
-end;
-
------------------------------------
--- onEventUpdate
------------------------------------
-
-function onEventUpdate(player,csid,option)
-    -- printf("CSID: %u",csid);
-    -- printf("RESULT: %u",option);
-end;
-
------------------------------------
--- onEventFinish
------------------------------------
-
-function onEventFinish(player,csid,option)
-    -- printf("CSID: %u",csid);
-    -- printf("RESULT: %u",option);
-end;
