+++ sql/mob_pools.sql
@@ -3690,7 +3690,7 @@ INSERT INTO `mob_pools` VALUES (3629,'Simorg','Simorg',265,0x0000C50800000000000
 INSERT INTO `mob_pools` VALUES (3630,'Simurgh','Simurgh',125,0x0000500100000000000000000000000000000000,1,10,5,240,125,0,1,0,0,2,0,32,0,3,0,0,45,1,0,125);
 INSERT INTO `mob_pools` VALUES (3631,'Sinister_Seidel','Sinister_Seidel',175,0x00009C0100000000000000000000000000000000,5,5,11,240,100,0,0,0,0,0,0,0,0,0,0,0,36,1,0,175);
 INSERT INTO `mob_pools` VALUES (3632,'Sippoy','Sippoy',450,0x0000A40800000000000000000000000000000000,2,5,12,360,100,0,0,0,0,2,0,0,0,0,0,0,158,1,0,450);
-INSERT INTO `mob_pools` VALUES (3633,'Siranpa-kamuy','Siranpa-kamuy',0,0x00008B0100000000000000000000000000000000,4,4,12,240,100,0,0,0,0,2,0,0,0,0,0,0,172,1,0,853);
+INSERT INTO `mob_pools` VALUES (3633,'Siranpa-kamuy','Siranpa-kamuy',245,0x00008B0100000000000000000000000000000000,4,4,12,240,100,0,0,0,0,2,0,0,0,0,0,0,172,1,0,853);
 INSERT INTO `mob_pools` VALUES (3634,'Sirin','Sirin',55,0x0000BC0100000000000000000000000000000000,1,1,5,240,100,0,1,1,1,16,0,0,7503,643,0,0,0,1,0,55);
 INSERT INTO `mob_pools` VALUES (3635,'Sirrush','Sirrush',174,0x0000490100000000000000000000000000000000,1,1,6,200,100,0,0,0,1,2,0,0,0,0,0,0,0,1,0,905);
 INSERT INTO `mob_pools` VALUES (3636,'Sisyphus','Sisyphus',135,0x0000B00100000000000000000000000000000000,1,1,6,240,100,0,0,0,0,2,0,0,0,0,0,0,0,1,0,135);
